LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-input-element/datetime-expected.txt

Date and Time Inputs


FAIL date type support on input element assert_equals: expected "date" but got "text"
PASS [date] The value must be a valid global date and time string
PASS [date] The min attribute must have a value that is a valid global date and time string
PASS [date] The max attribute must have a value that is a valid global date and time string
PASS [date] The step attribute must be expressed in seconds
PASS [date] stepUp method support on input 'date' element
PASS [date] stepDown method support on input 'date' element
FAIL [time] time type support on input element assert_equals: expected "time" but got "text"
PASS [time] The value must be a valid global date and time string
PASS [time] The min attribute must have a value that is a valid global date and time string
PASS [time] The max attribute must have a value that is a valid global date and time string
PASS [time] The step attribute must be expressed in seconds
PASS [time] stepUp method support on input 'time' element
PASS [time] stepDown method support on input 'time' element
FAIL datetime-local type support on input element assert_equals: expected "datetime-local" but got "text"
PASS [datetime-local] The must be a valid local date and time string
PASS [datetime-local] The min attribute must have a value that is a valid local date and time string
PASS [datetime-local] The max attribute must have a value that is a valid local date and time string
PASS [datetime-local] The step attribute must be expressed in seconds
PASS [datetime-local] stepUp method support on input 'datetime-local' element
PASS [datetime-local] stepDown method support on input 'datetime-local' element
FAIL month type support on input element assert_equals: expected "month" but got "text"
PASS [month] The value must be a value that is a valid global date and time string
PASS [month] The min attribute must have a value that is a valid global date and time string
PASS [month] The max attribute must have a value that is a valid global date and time string
PASS [month] The step attribute must be expressed in seconds
PASS [month] stepUp method support on input 'month' element
PASS [month] stepDown method support on input 'month' element
FAIL week type support on input element assert_equals: expected "week" but got "text"
PASS [week] The value must be a value that is a valid global date and time string
PASS [week] The min attribute must have a value that is a valid global date and time string
PASS [week] The max attribute must have a value that is a valid global date and time string
PASS [week] The step attribute must be expressed in seconds
PASS [week] stepUp method support on input 'week' element
PASS [week] stepDown method support on input 'week' element
